+++ Aero.Application/DTOs/ProcedureDto.cs
@@ -2,4 +2,4 @@ using Aero.Domain.Entities;
 
 namespace Aero.Application.DTOs;
 
-public sealed record ProcedureDto(int Id,int DeviceId,short DriverId, string Name,int TriggerId, List<ActionDto> Actions,int LocationId,bool IsActive) : BaseDto(LocationId,IsActive);
+public sealed record ProcedureDto(int Id,short DeviceId,short DriverId, string Name,int TriggerId, List<ActionDto> Actions,int LocationId,bool IsActive) : BaseDto(LocationId,IsActive);
