+++ Aero.Infrastructure/Persistences/Entities/DeviceCredential.cs
@@ -6,7 +6,7 @@ namespace Aero.Infrastructure.Persistences.Entities
     {
         [Key]
         public int id { get; set; }
-        public int device_id { get; set; } 
+        public short device_id { get; set; } 
         public short credential_id { get; set; }
         public Device hardware { get; set; }
         public Credential credential { get; set; }
