+++ Aero.Infrastructure/Persistences/Entities/SubFeature.cs
@@ -9,7 +9,7 @@ namespace Aero.Infrastructure.Persistences.Entities
         public int id { get; set; }
         public string name { get; set; } = string.Empty;
         public string path { get; set; } = string.Empty;
-        public short feature_id { get; set; }
+        public int feature_id { get; set; }
         public Feature feature { get; set; }
     }
 }
